+++ js/packages/web/src/views/packCreate/components/Sidebar/data.ts
@@ -5,15 +5,15 @@ export const STEPS_TITLES: Record<CreatePackSteps, string> = {
   [CreatePackSteps.SelectVoucher]: 'Select Voucher',
   [CreatePackSteps.AdjustQuantities]: 'Adjust Quantities',
   // [CreatePackSteps.SalesSettings]: 'Sales Settings',
-  [CreatePackSteps.DesignAndInfo]: 'Info & Design',
+  // [CreatePackSteps.DesignAndInfo]: 'Info & Design',
   [CreatePackSteps.ReviewAndMint]: 'Review & Mint',
 };
 
 export const CONTINUE_TITLES: Record<CreatePackSteps, string> = {
   [CreatePackSteps.SelectItems]: 'Continue to Voucher',
   [CreatePackSteps.SelectVoucher]: 'Continue to Quantities',
-  [CreatePackSteps.AdjustQuantities]: 'Continue to Design',
+  [CreatePackSteps.AdjustQuantities]: 'Continue to Mint',
   // [CreatePackSteps.SalesSettings]: 'Continue to Design',
-  [CreatePackSteps.DesignAndInfo]: 'Continue to Mint',
+  // [CreatePackSteps.DesignAndInfo]: 'Continue to Mint',
   [CreatePackSteps.ReviewAndMint]: 'Confirm & Mint',
 };